Added the `makeRequestObject` method for easy filling and standardization of the request object.
Added the `async_keep` global variable to accommodate multiple requests make in quick succession.

Added the ability to pass the URL to the callback along with the response for the same reasons `async_keep` was created. That's enabled by the global `send_url_to_callback` and can be passed along with each request in the same way as `verbose`.

Changed the way `verbose` triggers logging. Previously, the work of concatenating the message and calling the `log` function all took place for every call to `log` irrespective of the state of `verbose`. Now, `verbose` is checked before doing all that. This will greatly reduce the amount of work involved with logging, albeit at the expense of longer function bodies.